Add category and menu item video playback support
- Add videoId and videoFilename fields to MenuCategory
- Parse category video attributes from XML
- Implement category video playback in kiosk overlay
- Play category-specific videos when category is selected
- Auto-play initial category video on kiosk load
- Fix type conversion issue in idle_screen.dart (num to double)

+++ flutter_downloader/lib/screens/idle_screen.dart
@@ -158,8 +158,8 @@ class _IdleScreenState extends State<IdleScreen> {
                   child: FittedBox(
                     fit: BoxFit.cover,
                     child: SizedBox(
-                      width: _controller!.player.state.width ?? 1920,
-                      height: _controller!.player.state.height ?? 1080,
+                      width: (_controller!.player.state.width ?? 1920).toDouble(),
+                      height: (_controller!.player.state.height ?? 1080).toDouble(),
                       child: media_kit_video.Video(
                         controller: _controller!,
                         controls: media_kit_video.NoVideoControls,
